services/video-studio-agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
anthropic==0.25.0
litellm==1.26.0
google-cloud-storage==2.14.0
dropbox==11.36.0
azure-storage-blob==12.19.0
httpx==0.25.0
python-telegram-bot==20.7
twilio==8.11.0
supabase==2.3.0
redis==5.0.0
celery==5.3.4
opencv-python==4.8.1.78
Pillow==10.1.0
ffmpeg-python==0.2.1
huggingface-hub==0.19.4
requests==2.31.0
pytest==7.4.0
pytest-asyncio==0.21.0
black==23.12.0
pylint==3.0.0
